feat(Binding): Implement simple kafka binding.
- Clear unused and unecessary classes.
- Implement simple kafka binding with brokers and zookeepers.
