+++ .gitignore
@@ -35,3 +35,4 @@ osb-service/src/main/resources/static/
 osb-service/src/main/resources/templates/
 osb-service/osb-kafka.iml
 
+osb-service/src/main/resources/application-local.yml
